refactor enrolment serializers
